format(suggestions): refactor
- to split code logically and move util functions into util.go
- add comments for what each function does
